Intermediate. Broken. Some refactoring here. Move url handling to utils, rename reducer, remove cruft.
